mongoose и вся хурма с подключением к DB
