+++ config.js
@@ -1,3 +1,4 @@
 module.exports = {
-    PORT = process.env.PORT || 3000
-} 
+    PORT: process.env.PORT || 3000,
+    MONGO_URL: 'mongodb://localhost/blog'
+};
