+++ index.js
@@ -1,25 +1,12 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const app = express();
+const app = require('./app');
+const database = require('./database');
+const config = require('./config');
 
-app.set('view engine', 'ejs');
-app.use(bodyParser.urlencoded({extended: true}));
-
-
-const arr = [
-    'hello',
-    'world',
-    'test'
-];
-
-app.get('/', (req, res) => res.render('index', {arr: arr}));
-
-app.get('/create', (req, res) => res.render('create'));
-app.post('/create', (req, res) => {
-    arr.push(req.body.text)
-    res.redirect('/')
+database().then(info => {
+    console.log(`Connected to ${info.host}:${info.port}/${info.name}`)
+    app.listen(config.PORT, () => console.log(`Example app listening on port ${config.PORT}`));
+}).catch(() => {
+    console.error('Unable to connectn database');
+    process.exit(1);
 });
-
-
-
-app.listen(3000, () => console.log('Example app listening on port 3000!!!'));
+
