+++ package.json
@@ -7,7 +7,8 @@
     "body-parser": "^1.18.3",
     "cross-env": "^5.2.0",
     "ejs": "^2.6.1",
-    "express": "^4.16.4"
+    "express": "^4.16.4",
+    "mongoose": "^4.13.18"
   },
   "devDependencies": {
     "nodemon": "^1.18.9"
